refactor: 重构webview代码并优化UI样式
- 提取共享的webview模板到webviewTemplate.ts
- 将ChatListViewProvider分离到chatListView.ts
- 将ChatListPanel分离到chatListPanel.ts
- 修复灰matter模块缺失导致的扩展激活错误
- 优化.vscodeignore配置，缩小扩展包体积
- 添加esbuild配置进行代码打包
- 调整UI样式：筛选按钮文字变暗、标题悬停显示下划线
- 重排序聊天项操作按钮顺序
