+++ .myLastChat/last_chat.md
@@ -1,0 +1,201 @@
+---
+title: My Last Chat v0.0.1 项目总结
+description: 完成了类型图标、时间刷新、排序修复和 GitHub Release 发布
+project: My Last Chat
+type: planning
+createdAt: 2026-01-24T00:00:00Z
+updatedAt: 2026-01-24T00:00:00Z
+tags: [项目完成, 扩展开发, GitHub发布]
+---
+
+# My Last Chat v0.0.1 项目总结
+
+## 概述
+完成了 VS Code 扩展 "My Last Chat" v0.0.1 版本的最后阶段工作，包括功能完善、Bug 修复和 Release 发布。
+
+## 本次工作内容
+
+### 1. 类型图标功能
+为聊天摘要项添加基于 `type` 字段的不同图标：
+- `coding` → 代码图标 (codicon-code)
+- `debug` → 调试图标 (codicon-bug)
+- `planning` → 计划图标 (codicon-checklist)
+- `spec` → 文档图标 (codicon-file-text)
+- 默认 → 讨论图标 (codicon-comment-discussion)
+
+在两个视图（侧边栏和编辑器面板）中都添加了 CSS 样式和 JavaScript 函数：
+
+```typescript
+function getTypeIcon(type) {
+  switch (type) {
+    case 'coding': return 'codicon-code';
+    case 'debug': return 'codicon-bug';
+    case 'planning': return 'codicon-checklist';
+    case 'spec': return 'codicon-file-text';
+    default: return 'codicon-comment-discussion';
+  }
+}
+```
+
+### 2. 实时时间刷新
+添加了时间自动刷新功能，每分钟更新一次时间显示，保持列表时间的实时性：
+
+```typescript
+let currentItems = [];
+
+function updateTimes() {
+  document.querySelectorAll('.chat-item').forEach(item => {
+    const path = item.dataset.path;
+    const it = currentItems.find(i => i.filePath === path);
+    if (it) {
+      const timeEl = item.querySelector('.item-time');
+      if (timeEl) {
+        timeEl.textContent = formatTime(it.updatedAt || it.createdAt);
+      }
+    }
+  });
+}
+
+setInterval(updateTimes, 60000);
+```
+
+### 3. 默认排序修复
+修复了 `refresh()` 方法没有应用排序的问题。现在刷新时默认按最近更新时间降序排列：
+
+```typescript
+public async refresh(): Promise<void> {
+  let summaries = await this._storageService.getAllChatSummaries();
+  
+  // 默认按最近更新时间降序排序
+  summaries.sort((a, b) => {
+    const aTime = a.metadata.updatedAt || a.metadata.createdAt || '';
+    const bTime = b.metadata.updatedAt || b.metadata.createdAt || '';
+    return new Date(bTime).getTime() - new Date(aTime).getTime();
+  });
+  
+  // ... 发送数据
+}
+```
+
+在两个视图（ChatListViewProvider 和 ChatListPanel）中都进行了同样的修复。
+
+### 4. 编辑器面板状态监听
+添加了面板状态变化监听，当面板从隐藏状态重新显示时，自动刷新列表数据：
+
+```typescript
+this._panel.onDidChangeViewState(
+  async (e) => {
+    if (e.webviewPanel.visible) {
+      await this._refresh();
+    }
+  },
+  null,
+  this._disposables
+);
+```
+
+## 文档和发布
+
+### README 更新
+更新了 README.md 文档，添加了：
+- 新功能说明（时间分组、类型图标、实时刷新、收藏功能）
+- Copilot 工具参数表格
+- scope 参数说明
+- 命令列表表格化
+- 更新日志部分
+
+### GitHub 代理配置
+解决了 Git 终端连接 GitHub 的问题，配置代理：
+```bash
+git config --global http.proxy http://127.0.0.1:7890
+git config --global https.proxy http://127.0.0.1:7890
+```
+
+### Release 发布
+1. 创建 Git tag: `git tag -a v0.0.1 -m "Release v0.0.1 - 初始版本"`
+2. 安装 GitHub CLI: `winget install --id GitHub.cli`
+3. 发布 Release 到 GitHub，包含：
+   - `my-last-chat-0.0.1.vsix` 安装包
+   - 详细的功能说明和安装指导
+
+**Release 地址**: https://github.com/isWittHere/my-last-chat/releases/tag/v0.0.1
+
+## 技术要点
+
+### 使用的技术栈
+- **VS Code Extension** - TypeScript、WebView
+- **@vscode/codicons** - 图标库
+- **gray-matter** - Markdown frontmatter 解析
+- **languageModelTools** - Copilot 工具集成
+- **Git & GitHub CLI** - 版本管理和发布
+
+### 文件修改概览
+
+#### src/webview.ts
+- 添加 `getTypeIcon()` 函数（两个视图）
+- 添加 `currentItems` 数组和 `updateTimes()` 函数（两个视图）
+- 设置 `setInterval(updateTimes, 60000)` 定时器（两个视图）
+- 更新 `render()` 函数，添加图标显示和 `currentItems` 赋值（两个视图）
+- 修复 `refresh()` 和 `_refresh()` 方法，添加默认排序逻辑
+- ChatListPanel 构造函数中添加 `onDidChangeViewState` 监听器
+
+#### README.md
+- 更新功能特性部分
+- 添加 Copilot 工具说明表格
+- 添加 scope 参数文档
+- 更新命令列表为表格格式
+- 添加更新日志版本 v0.0.1
+
+## 问题解决
+
+### 问题 1: Git 推送失败
+**症状**: 终端无法连接 GitHub，但浏览器可以访问
+**原因**: 使用 Clash 代理，终端未配置代理
+**解决方案**: 配置全局 Git 代理
+```bash
+git config --global http.proxy http://127.0.0.1:7890
+git config --global https.proxy http://127.0.0.1:7890
+```
+
+### 问题 2: 刷新后列表不按默认排序
+**症状**: 虽然排序下拉框默认选中"最近更新"，但列表不按时间排序
+**原因**: `refresh()` 方法直接返回数据库数据，没有应用排序
+**解决方案**: 在 `refresh()` 和 `_refresh()` 中添加排序逻辑
+
+### 问题 3: 编辑器面板数据不刷新
+**症状**: 从其他标签页切换回编辑器面板时，列表不自动刷新
+**原因**: 缺少面板状态变化事件监听
+**解决方案**: 添加 `onDidChangeViewState` 事件监听器
+
+## Commits 记录
+
+1. **61fa8f1** - feat: 添加类型图标、实时时间刷新、默认排序功能
+   - 为不同类型的聊天摘要添加对应的图标
+   - 添加时间实时刷新功能 (每分钟更新)
+   - 修复 refresh() 方法，添加默认按最近更新时间降序排序
+   - 添加编辑器面板状态监听，切换回面板时自动刷新数据
+
+2. **b24db14** - docs: 更新 README 文档
+   - 更新功能特性描述，添加新功能说明
+   - 添加 Copilot 工具参数表格
+   - 添加 scope 参数说明
+   - 更新命令列表为表格格式
+   - 添加更新日志
+
+## 后续改进方向
+
+1. **发布到 VS Code Marketplace** - 使 非正式安装更便捷
+2. **类型图标自定义** - 允许用户自定义类型和对应的图标
+3. **导入/导出功能** - 支持聊天摘要的备份和迁移
+4. **高级搜索** - 支持正则表达式和日期范围搜索
+5. **性能优化** - 虚拟列表优化大量数据的显示
+
+## 总结
+
+v0.0.1 版本成功完成，扩展功能完整、稳定，已发布到 GitHub Releases。用户可以通过以下方式安装：
+
+1. **从 GitHub Release 下载** vsix 文件
+2. **VS Code 命令面板** - Extensions: Install from VSIX
+3. **终端安装** - `code --install-extension my-last-chat-0.0.1.vsix --force`
+
+扩展现已可供使用，具备聊天摘要管理、Copilot 集成、多视图显示等完整功能。
