+++ .vscodeignore
@@ -1,1 +1,12 @@
-.vscode-test
+.vscode/**
+.vscode-test/**
+src/**
+.git/**
+.gitignore
+.myLastChat/**
+*.vsix
+tsconfig.json
+esbuild.js
+**/*.ts
+**/*.map
+node_modules/**
